requirements.txt
#python==3.6.15
absl-py==0.9.0
#amqp==2.6.1
#APScheduler==3.6.3
#arrow==0.14.7
asgiref==3.2.3
astor==0.8.1
astroid==2.4.1
bcrypt==3.1.7
billiard==3.6.3.0
#boto3==1.12.22
#botocore==1.15.22
#cachetools==4.1.1
#celery==4.4.7
#celery-progress==0.0.12
#certifi==2019.11.28
#cffi==1.14.1
#chardet==3.0.4
#colorama==0.4.3
#contextlib2==21.6.0
#crypto==1.4.1
#cryptography==3.0
cycler==0.10.0
Cython==0.29.24
#dataclasses==0.8
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.1
#django-admin-list-filter-dropdown==1.0.3
#django-celery-results==1.2.1
#django-cors-headers==3.2.1
#django-extra-fields==3.0.2
django-reset-migrations==0.4.0
django-storages==1.9.1
djangorestframework==3.12.4
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
gast==0.5.2
gcloud==0.17.0
geographiclib==1.50
geopy==1.21.0
#google-auth==1.4.2
#google-colab==1.0.0
#google-pasta==0.2.0
#googleapis-common-protos==1.52.0
grpcio==1.40.0
gunicorn==19.9.0
h5py==2.10.0
httplib2==0.18.1
icalendar==4.0.7
ics==0.7
idna==2.8
imageio==2.8.0
importlib-metadata==4.8.1
install==1.3.3
#ipykernel==5.3.4
#ipython==7.16.1
#isort==4.3.21
jmespath==0.9.5
joblib==0.15.1
jws==0.1.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
#kiwisolver==1.2.0
#kombu==4.6.11
lazy-object-proxy==1.4.3
lxml==4.6.3
Markdown==3.3.4
#matplotlib==3.2.2
mccabe==0.6.1
mlxtend==0.17.2
#Naked==0.1.31N
nest-asyncio==1.4.1
networkx==2.4
numpy==1.19.5
oauth2client==3.0.0
opencv-python==4.5.3.56
pandas==0.25.3
pem==20.1.0
pika==1.1.0
Pillow==8.3.2
portpicker==1.2.0
prompt-toolkit==3.0.8
protobuf==3.17.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.4.3
pydicom==1.4.2
#pyfcm==1.4.7 #n
pyflakes==2.1.1
pylint==2.5.2
pyparsing==2.4.7
#Pyrebase==3.0.27 #n
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
#requests==2.21.0
requests-toolbelt==0.7.0
rsa==4.6
#s3transfer==0.3.3
#scikit-image==0.17.2
scikit-learn==0.23.1
#scipy==1.4.1
#seaborn==0.10.0
#Shapely==1.7.1
#shellescape==3.8.1
#simplegeneric==0.8.1
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
TatSu==4.4.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.25.7
vine==1.3.0
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
#xgboost==1.4.2
zipp==3.5.0

